#58 - 임시 컨트롤러 삭제 및 본 컨트롤러 연동
- todo
1) 팔로우중인 유저 상단 배치
2) 테스트 케이스 작성
