+++ src/test/java/com/capstone/pick/controller/PickControllerTest.java
@@ -3,6 +3,7 @@ package com.capstone.pick.controller;
 import com.capstone.pick.config.TestSecurityConfig;
 import com.capstone.pick.controller.request.PickRequest;
 import com.capstone.pick.service.PickService;
+import com.capstone.pick.service.UserService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -34,6 +35,8 @@ public class PickControllerTest {
     private final ObjectMapper objectMapper;
 
     @MockBean private PickService pickService;
+
+    @MockBean private UserService userService;
 
     public PickControllerTest(@Autowired MockMvc mvc, @Autowired ObjectMapper objectMapper) {
         this.mvc = mvc;
